requirements.txt
codecov==2.0.5
coverage==4.2
Django==1.10.1
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
python-dateutil==2.5.3
ovp-projects==0.1.0
